refactor(server): centralize style source and provider normalization
